+++ apps/web/src/app/app.tsx
@@ -48,15 +48,18 @@ export function App() {
       theme={value.theme}
       onChange={(value) => {
         localforage.setItem(MAIN_BOARD_CONTENT_KEY, value);
+        setValue(value);
       }}
       afterInit={(board) => {
         console.log('board initialized');
+        /*
         console.log(
           `add __drawnix__web__debug_log to window, so you can call add log anywhere, like: window.__drawnix__web__console('some thing')`
         );
         (window as any)['__drawnix__web__console'] = (value: string) => {
           addDebugLog(board, value);
         };
+        */
       }}
     ></Drawnix>
   );