+++ packages/drawnix/src/plugins/with-hotkey.ts
@@ -1,4 +1,9 @@
-import { BoardTransforms, PlaitBoard, PlaitPointerType } from '@plait/core';
+import {
+  BoardTransforms,
+  getSelectedElements,
+  PlaitBoard,
+  PlaitPointerType,
+} from '@plait/core';
 import { isHotkey } from 'is-hotkey';
 import { addImage, saveAsImage } from '../utils/image';
 import { saveAsJSON } from '../data/json';
@@ -12,11 +17,16 @@ export const buildDrawnixHotkeyPlugin = (
   updateAppState: (appState: Partial<DrawnixState>) => void
 ) => {
   const withDrawnixHotkey = (board: PlaitBoard) => {
-    const { globalKeyDown } = board;
+    const { globalKeyDown, keyDown } = board;
     board.globalKeyDown = (event: KeyboardEvent) => {
+      const isTypingNormal =
+        event.target instanceof HTMLInputElement ||
+        event.target instanceof HTMLTextAreaElement;
       if (
-        PlaitBoard.getMovingPointInBoard(board) ||
-        PlaitBoard.isMovingPointInBoard(board)
+        !isTypingNormal &&
+        (PlaitBoard.getMovingPointInBoard(board) ||
+          PlaitBoard.isMovingPointInBoard(board)) &&
+        !PlaitBoard.hasBeenTextEditing(board)
       ) {
         if (isHotkey(['mod+shift+e'], { byKey: true })(event)) {
           saveAsImage(board, true);
@@ -38,44 +48,79 @@ export const buildDrawnixHotkeyPlugin = (
           event.preventDefault();
           return;
         }
-        if (event.key === 'h') {
-          BoardTransforms.updatePointerType(board, PlaitPointerType.hand);
-          updateAppState({ pointer: PlaitPointerType.hand });
-        }
-        if (event.key === 'v') {
-          BoardTransforms.updatePointerType(board, PlaitPointerType.selection);
-          updateAppState({ pointer: PlaitPointerType.selection });
-        }
-        if (event.key === 'm') {
-          setCreationMode(board, BoardCreationMode.dnd);
-          BoardTransforms.updatePointerType(board, MindPointerType.mind);
-          updateAppState({ pointer: MindPointerType.mind });
-        }
-        if (event.key === 'p') {
-          setCreationMode(board, BoardCreationMode.drawing);
-          BoardTransforms.updatePointerType(board, FreehandShape.feltTipPen);
-          updateAppState({ pointer: FreehandShape.feltTipPen });
-        }
-        if (event.key === 'a') {
-          setCreationMode(board, BoardCreationMode.drawing);
-          BoardTransforms.updatePointerType(board, ArrowLineShape.straight);
-          updateAppState({ pointer: ArrowLineShape.straight });
-        }
-        if (event.key === 'r' || event.key === 'o' || event.key === 't') {
-          const keyToPointer = {
-            r: BasicShapes.rectangle,
-            o: BasicShapes.ellipse,
-            t: BasicShapes.triangle,
-          };
-          setCreationMode(board, BoardCreationMode.drawing);
-          BoardTransforms.updatePointerType(board, keyToPointer[event.key]);
-          updateAppState({ pointer: keyToPointer[event.key] });
-        }
         if (isHotkey(['mod+u'])(event)) {
           addImage(board);
         }
+        if (!event.altKey && !event.metaKey && !event.ctrlKey) {
+          if (event.key === 'h') {
+            BoardTransforms.updatePointerType(board, PlaitPointerType.hand);
+            updateAppState({ pointer: PlaitPointerType.hand });
+          }
+          if (event.key === 'v') {
+            BoardTransforms.updatePointerType(
+              board,
+              PlaitPointerType.selection
+            );
+            updateAppState({ pointer: PlaitPointerType.selection });
+          }
+          if (event.key === 'm') {
+            setCreationMode(board, BoardCreationMode.dnd);
+            BoardTransforms.updatePointerType(board, MindPointerType.mind);
+            updateAppState({ pointer: MindPointerType.mind });
+          }
+          if (event.key === 'e') {
+            setCreationMode(board, BoardCreationMode.drawing);
+            BoardTransforms.updatePointerType(board, FreehandShape.eraser);
+            updateAppState({ pointer: FreehandShape.eraser });
+          }
+          if (event.key === 'p') {
+            setCreationMode(board, BoardCreationMode.drawing);
+            BoardTransforms.updatePointerType(board, FreehandShape.feltTipPen);
+            updateAppState({ pointer: FreehandShape.feltTipPen });
+          }
+          if (event.key === 'a' && !isHotkey(['mod+a'])(event)) {
+            // will trigger editing text
+            if (getSelectedElements(board).length === 0) {
+              setCreationMode(board, BoardCreationMode.drawing);
+              BoardTransforms.updatePointerType(board, ArrowLineShape.straight);
+              updateAppState({ pointer: ArrowLineShape.straight });
+            }
+          }
+          if (event.key === 'r' || event.key === 'o' || event.key === 't') {
+            const keyToPointer = {
+              r: BasicShapes.rectangle,
+              o: BasicShapes.ellipse,
+              t: BasicShapes.text,
+            };
+            if (keyToPointer[event.key] === BasicShapes.text) {
+              setCreationMode(board, BoardCreationMode.dnd);
+            } else {
+              setCreationMode(board, BoardCreationMode.drawing);
+            }
+            BoardTransforms.updatePointerType(board, keyToPointer[event.key]);
+            updateAppState({ pointer: keyToPointer[event.key] });
+          }
+          event.preventDefault();
+          return;
+        }
       }
       globalKeyDown(event);
+    };
+
+    board.keyDown = (event: KeyboardEvent) => {
+      if (isHotkey(['mod+z'], { byKey: true })(event)) {
+        board.undo();
+        event.preventDefault();
+        return;
+      }
+
+      if (isHotkey(['mod+shift+z'], { byKey: true })(event)) {
+        board.redo();
+        event.preventDefault();
+        return;
+      }
+
+      keyDown(event);
     };
 
     return board;